requirements.txt
altgraph==0.17.5
colorama==0.4.6
packaging==25.0
pillow==12.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
setuptools==80.9.0
